Still working on Pacman moving

+++ src/main/mapComponents/Coordinate.java
@@ -1,29 +1,29 @@
 package main.mapComponents;
 
 public class Coordinate {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
     // 0,0 is top left like graph
 
-    public Coordinate(int x, int y) {
+    public Coordinate(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 }
